server/requirements.txt
django>=4.2.0
djangorestframework>=3.14.0
torch>=2.0.0
torchaudio>=2.0.0
librosa>=0.10.0
transformers>=4.30.0
numpy>=1.24.0
scikit-learn>=1.3.0